refactor #8 피드백 반영 구조 조정
