+++ src/pages/main/ui/MainIntro.tsx
