+++ src/pages/main/ui/RoomCard.tsx
@@ -1,16 +1,25 @@
 import { Clock3, MapPin, Users } from 'lucide-react';
 
-import type { Room } from '../../../entities/room/model/types';
+import type { Room } from '@/entities/room/model/types';
 
 interface RoomCardProps {
   room: Room;
 }
 
+const accentColorClassNameMap = {
+  blue: 'bg-blue-500',
+  indigo: 'bg-indigo-500',
+  pink: 'bg-pink-500',
+} as const;
+
 const RoomCard = ({ room }: RoomCardProps) => {
+  const progress = (room.currentCount / room.maxCount) * 100;
+  const accentColorClassName = accentColorClassNameMap[room.accentColor];
+
   return (
     <article className="overflow-hidden rounded-[24px] border border-slate-200 bg-white shadow-[0_8px_24px_rgba(15,23,42,0.04)]">
       <div className="flex items-start gap-3 p-5">
-        <div className={`h-full w-1.5 self-stretch rounded-full ${room.accentClassName}`} />
+        <div className={`h-full w-1.5 self-stretch rounded-full ${accentColorClassName}`} />
         <div className="flex-1">
           <h3 className="text-base font-semibold text-slate-900">{room.title}</h3>
           <div className="mt-2 flex flex-wrap items-center gap-2 text-xs text-slate-500">
@@ -42,8 +51,8 @@ const RoomCard = ({ room }: RoomCardProps) => {
 
           <div className="mt-4 h-1.5 w-full overflow-hidden rounded-full bg-slate-100">
             <div
-              className={`h-full rounded-full ${room.accentClassName}`}
-              style={{ width: `${room.progress}%` }}
+              className={`h-full rounded-full ${accentColorClassName}`}
+              style={{ width: `${progress}%` }}
             />
           </div>
