+++ src/pages/main/ui/RoomStatus.tsx
